Start tensorboard from script * separate evaluation result writers * mitigate excessive logging of tensorboard * add unit tests
